404, 500, and bootstrap nav-bar
django will use the 400.html, and the 500.html based
on the location of the files, no other code required.
for a more complicated view you will need to set;
handler404 = 'views.custom_404_view' in your URLconf.
the same will apply to the 500, 403, and 400 code
I added the bootstrap navbar (fixed), and all the css
classes in the style_sheet used by the navbar (as far as i know).